forward declare CmdEvent, CmdResultError and CmdResult
move conversions into its own file
